requirements.txt
# PolyLens Core - Blockchain & Data
web3>=6.0.0
eth-abi>=4.0.0
eth-utils>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=2.0.0

# API Server & WebSocket (FastAPI)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0
websockets>=11.0

# Task Scheduling
apscheduler>=3.10.0

# CLI & Utilities
click>=8.0.0
tqdm>=4.65.0

# Legacy Dashboard (Optional/Streamlit fallback)
streamlit>=1.28.0
plotly>=5.18.0
streamlit-autorefresh>=1.0.0